add option to download all of /suppl

+++ download_from_geo.sh
@@ -3,20 +3,21 @@
 # downloads data for all accessions in an input file
 # each accession gets its own folder
 
-infile=  expand=false  decompress=false  cleanup=false
+infile=  expand=false  decompress=false  cleanup=false  all=false
 
 usage="$(basename "$0") -- download data from NCBI GEO
   -h  show help and exit
   -f [file]   input file with a list of GEO accession names
   -e          expand downloaded archives
-  -c          cleanup (remove tar files if expanded)"
+  -c          cleanup (remove tar files if expanded)
+  -a          download all files (not just RAW.tar)"
 
 if [ $# -eq 0 ]; then
   printf "$usage\n"
   exit 1
 fi
 
-while getopts ":f:ech" opt; do
+while getopts ":f:ecah" opt; do
   case $opt in
     h)  printf "$usage\n"
         exit
@@ -29,6 +30,8 @@ while getopts ":f:ech" opt; do
         ;;
     c)  cleanup=true
         ;;
+    a)  all=true
+	;;
     :)  printf "missing argument for -%s\n" "$OPTARG" >&2
         printf "$usage\n" >&2
         exit 1
@@ -52,12 +55,16 @@ while read acc; do
 	printf "Downloading ${acc} from GEO\n"
 	prefix=${acc::${#acc}-3}
     mkdir $acc
-    curl "ftp://ftp.ncbi.nlm.nih.gov/geo/series/${prefix}nnn/${acc}/suppl/${acc}_RAW.tar" -o $acc/${acc}_RAW.tar
-    if [ $expand == true ]; then
-      printf "Expanding ${acc}"
-      tar -xf $acc/${acc}_RAW.tar -C $acc
-      if [ $cleanup == true ]; then
-        rm $acc/${acc}_RAW.tar
+    if [ $all == true ]; then
+      wget -r --no-parent --cut-dirs=6 --directory-prefix=$acc "ftp://ftp.ncbi.nlm.nih.gov/geo/series/${prefix}nnn/${acc}/suppl"
+    else
+      curl "ftp://ftp.ncbi.nlm.nih.gov/geo/series/${prefix}nnn/${acc}/suppl/${acc}_RAW.tar" -o $acc/${acc}_RAW.tar
+      if [ $expand == true ]; then
+        printf "Expanding ${acc}"
+        tar -xf $acc/${acc}_RAW.tar -C $acc
+        if [ $cleanup == true ]; then
+          rm $acc/${acc}_RAW.tar
+        fi
       fi
-	fi
+    fi
 done < $infile
